split modules into two separate files
